primer Commit

+++ Repositories/PresupuestosRepository.cs
@@ -5,11 +5,25 @@ using tl2_tp8_2025_BautistaAlvarez.Models;
 using tl2_tp8_2025_BautistaAlvarez.Interfaces;
 public class PresupuestosRepository : IPresupuestosRepository
 {
-    string cadenaConexion = "Data Source=DB/Tienda.db"; //conexion para todo el repositorio
-
+    //string cadenaConexion = "Data Source=DB/Tienda.db"; //conexion para todo el repositorio
+    //tp11
+    private readonly string _ConnectionString;//cambio a esto para la inyeccion de dependencia
+    //creo constructor para la inyeccion
+    public PresupuestosRepository(string connectionString)
+    {
+        _ConnectionString = connectionString;//luego tuve que cambiar cada cadena conexon por _ConnectionString
+    }
     public void CrearPresupuesto(Presupuestos presupuesto)
     {
-        using var conexion = new SqliteConnection(cadenaConexion);//usar using
+        //tp11
+        if (string.IsNullOrWhiteSpace(presupuesto.NombreDestinatario))
+            throw new Exception("No puede estar vacio el nombre del destinatario");
+        if(presupuesto.FechaCreacion == DateOnly.MinValue)//si el usuario no ingresa fecha, por defecto se ingresa la fecha minima entonces si eso pasa sale una excepcion
+            throw new Exception("Debe ingresar una fecha valida");
+        if(presupuesto.FechaCreacion == DateOnly.FromDateTime(DateTime.Now))//si el usuario ingresa una fecha futura
+            throw new Exception("Debe ingresar una fecha que no sea futura");
+        //tp11
+        using var conexion = new SqliteConnection(_ConnectionString);//usar using
         conexion.Open();//abro la conexion usando using para que se abra y cierre cuando sea deje de usarse
 
         string sql = "INSERT INTO Presupuestos (NombreDestinatario, FechaCreacion) VALUES (@NombreDestinatario, @FechaCreacion)";//codigo sql
@@ -23,14 +37,16 @@ public class PresupuestosRepository : IPresupuestosRepository
         //presupuesto.FechaCreacion.ToDateTime(TimeOnly.MinValue))) //otra opcion
         //como uso DateOnly debo pasar de dateOnly a DateTime con el comando.ToDateTime() y dentro del parentesis van los minutos que inicio por eso timeonly.minvalue
 
-        comando.ExecuteNonQuery();//ejecuto sino tengo que mostrar nada
+        int filasAfectadas = comando.ExecuteNonQuery(); //ejecuto sino tengo que mostrar nada y número de filas insertadas
+        if (filasAfectadas == 0)//sino afecto ninguna fila, daria 0 y en ese caso no se pudo crear presupuesto
+            throw new Exception("No se pudo crear el presupuesto, inténtelo de nuevo.");
     }
 
     public List<Presupuestos> ListarPresupuesto()
     {
         var listado = new List<Presupuestos>();//inicio lista
 
-        using var conexion = new SqliteConnection(cadenaConexion);
+        using var conexion = new SqliteConnection(_ConnectionString);
         conexion.Open();
 
         string sql = "SELECT * FROM Presupuestos";//* es para elegir todo los campos, sin where para elegir toda la tabla
@@ -83,13 +99,17 @@ public class PresupuestosRepository : IPresupuestosRepository
                 presupuesto.Detalle.Add(presupuestoDetalle);//lo agrego a la lista hasta que el lector termine de leer
             }//una vez que termine de leer todo lo relacionado con este presupuesto, el foreach pasa al siguiente presupuesto repitiendo el ciclo
         }
+        //tp11
+        if(listado.Count == 0)
+            throw new Exception("La lista de presupuesto esta vacia");
+
         return listado;//una vez terminado el foreach retorno el listado        
     }
     public Presupuestos PresupuestoPorId(int idPresupuesto)
     {
         var presupuesto = new Presupuestos();//creo una variable presupuesto para devolver
 
-        using var conexion = new SqliteConnection(cadenaConexion);
+        using var conexion = new SqliteConnection(_ConnectionString);
         conexion.Open();
 
         string sqlPresupuesto = "SELECT idPresupuesto, NombreDestinatario, FechaCreacion FROM Presupuestos WHERE idPresupuesto = @idPresupuesto";
@@ -103,6 +123,10 @@ public class PresupuestosRepository : IPresupuestosRepository
             presupuesto.IdPresupuesto = Convert.ToInt32(lectorPresupuesto["idPresupuesto"]); //convierto y agrego los datos encontrados
             presupuesto.NombreDestinatario = lectorPresupuesto["NombreDestinatario"].ToString();
             presupuesto.FechaCreacion = DateOnly.FromDateTime(Convert.ToDateTime(lectorPresupuesto["FechaCreacion"]));//converto a datetime por las dudas y luego uso date only.fromdate time para sacar solo la fecha
+        }
+        else//TP11
+        {
+            throw new Exception($"Presupuesto de id {idPresupuesto} inexistente");//si no encuentra el registro salta una excepcion
         }
         //lectorPresupuesto.Close(); aqui tendria que cerrar el lector pero como uso using ya lo hace solo
 
@@ -151,7 +175,15 @@ public class PresupuestosRepository : IPresupuestosRepository
     }
     public void AgregarProducto(int idPresupuesto, int idProducto, int cantidad)
     {
-        using var conexion = new SqliteConnection(cadenaConexion);
+        //tp11
+        if (idPresupuesto <= 0)
+            throw new Exception("El ID del presupuesto no es válido.");
+        if (idProducto <= 0)
+            throw new Exception("El ID del producto no es válido.");
+        if (cantidad <= 0)
+            throw new Exception("La cantidad debe ser mayor a cero.");
+        //tp11
+        using var conexion = new SqliteConnection(_ConnectionString);
         conexion.Open();
 
         string sql = "INSERT INTO PresupuestosDetalle(idPresupuesto, idProducto, Cantidad) VALUES (@idPresupuesto, @idProducto, @Cantidad)";
@@ -161,29 +193,38 @@ public class PresupuestosRepository : IPresupuestosRepository
         comando.Parameters.AddWithValue("@idProducto", idProducto);//una forma de agregar valores de manera rapida
         comando.Parameters.Add(new SqliteParameter("@Cantidad", cantidad));
 
-        comando.ExecuteNonQuery();
+        int filasAfectadas = comando.ExecuteNonQuery();//como no hay nada que mostrar ejecuto, a su vez guardo el numero de filas afectadas
+
+        if (filasAfectadas == 0)//si el numero de filas afectadas es 0 significa que no se efectuo por tanto lanza una excepcion
+            throw new Exception("No se pudo agregar el producto al presupuesto.");
     }
     public void EliminarPresupuestoPorId(int idPresupuesto)
     {
-        using var conexion = new SqliteConnection(cadenaConexion);
+        using var conexion = new SqliteConnection(_ConnectionString);
         conexion.Open();
         //primero debo borrar el presupuesto de la tabla hijo (PresupuestosDetalles) ya que contiene el idPresupuesto y esta la clave foragnea, luego elimino la tabla origen o principal
         string sqlDetalle = "DELETE FROM PresupuestosDetalle where idPresupuesto = @idPresupuesto";//tabla hijo o la que esta conectada con la clave foranea
 
         using var comandoDetalle = new SqliteCommand(sqlDetalle, conexion);
         comandoDetalle.Parameters.Add(new SqliteParameter("@idPresupuesto", idPresupuesto));
-        comandoDetalle.ExecuteNonQuery();
+        comandoDetalle.ExecuteNonQuery();//si el presupuesto o el detalle no existe no borra nada y no da error
         //tabla principal
         string sqlPrincipal = "DELETE FROM Presupuestos WHERE idPresupuesto = @idPresupuesto";
 
         using var comando = new SqliteCommand(sqlPrincipal, conexion);
 
-        comando.Parameters.Add(new SqliteParameter("@idPresupuesto", idPresupuesto));
-        comando.ExecuteNonQuery();
+        comando.Parameters.Add(new SqliteParameter("@idPresupuesto", idPresupuesto));//agrego el valor al sql
+        
+        //tp11
+        int filaAfectada = comando.ExecuteNonQuery();//ejecuto y guardo el numeros de filas, ya que execute da el numero de filas, en este caso siempre seria 1
+        if (filaAfectada == 0)//comando.ExecuteNonQuery() ejecuta y luego guardo el numero, si da 0 significa que no hizo nada entonces se hace la excepcion
+        {
+            throw new Exception($"Presupuesto de id: {idPresupuesto} inexistente o ya eliminado");
+        }
     }
     public void ModificarPresupuesto(int idPresupuesto, Presupuestos presupuesto)
     {
-        using var conexion = new SqliteConnection(cadenaConexion);
+        using var conexion = new SqliteConnection(_ConnectionString);
         conexion.Open();
 
         string sqlPresupuesto = "UPDATE Presupuestos SET NombreDestinatario = @NombreDestinatario WHERE idPresupuesto = @idPresupuesto";
@@ -191,7 +232,12 @@ public class PresupuestosRepository : IPresupuestosRepository
 
         comandoPresupuesto.Parameters.Add(new SqliteParameter("@NombreDestinatario", presupuesto.NombreDestinatario));
         comandoPresupuesto.Parameters.Add(new SqliteParameter("@idPresupuesto", idPresupuesto));
-        comandoPresupuesto.ExecuteNonQuery();
+
+        //tp11
+        int filasModificada = comandoPresupuesto.ExecuteNonQuery();
+        if(filasModificada == 0)
+            throw new Exception($"Presupuesto con ID {idPresupuesto} no existe y no puede ser modificado.");
+
         //borrar detalle
         string sqlBorrarDetalle = "DELETE FROM PresupuestosDetalle WHERE idPresupuesto = @idPresupuesto";
 
@@ -208,15 +254,16 @@ public class PresupuestosRepository : IPresupuestosRepository
             comandoDetalle.Parameters.Add(new SqliteParameter("@idPresupuesto", idPresupuesto));//va dentro ya que el comando se debe actualizar
             comandoDetalle.Parameters.Add(new SqliteParameter("@idProducto", detalle.Producto.IdProducto));
             comandoDetalle.Parameters.Add(new SqliteParameter("@Cantidad", detalle.Cantidad));
-            comandoDetalle.ExecuteNonQuery();
+            //tp11
+            int filas = comandoDetalle.ExecuteNonQuery();//ejecuto y guardo el numero filas
+            if (filas == 0)//si diera 0, significa que no se hizo bien el cambio
+                throw new Exception($"No se pudo agregar el producto de ID {detalle.Producto.IdProducto} al presupuesto de ID {idPresupuesto}.");
         }
-
-
     }
 
     public bool ExistePresupuesto(int idPresupuesto)
     {
-        using var conexion = new SqliteConnection(cadenaConexion);
+        using var conexion = new SqliteConnection(_ConnectionString);
         conexion.Open();
         //select COUNT(*) sirve para contar todas las filas de la talba presupuesto, y el WHERE lo uso para que solo cuente cuando haya coincidencia de ID. Lo cual siempre me devolvera 1 o 0
         string sql = "SELECT COUNT(*) FROM Presupuestos WHERE idPresupuesto = @idPresupuesto";
@@ -228,7 +275,7 @@ public class PresupuestosRepository : IPresupuestosRepository
     }
     public bool ExisteProducto(int idProducto)
     {
-        using var conexion = new SqliteConnection(cadenaConexion);
+        using var conexion = new SqliteConnection(_ConnectionString);
         conexion.Open();
         //select COUNT(*) sirve para contar todas las filas de la talba presupuesto, y el WHERE lo uso para que solo cuente cuando haya coincidencia de ID. Lo cual siempre me devolvera 1 o 0
         string sql = "SELECT COUNT(*) FROM Productos WHERE idProducto = @idProducto";
